Use 64-bit: numeric type fixes
Unfortunately a printf precision specifier (‘%.*s’) can not come with a width specifier so we have to cast to int. The width specifier ‘t’ is used for ptrdiff_t.
The int → NSInteger change fixed a bug with popup menu positioning, but there was no associated warning or error. It's possible there are more such bugs that we haven't found yet!

+++ Frameworks/OakAppKit/src/OakStatusBar.h
@@ -136,7 +136,7 @@ PUBLIC @interface OakStatusBar : OakControl
 @property (nonatomic, readonly) CGFloat minimumWidth;
 @property (nonatomic, assign) NSInteger borderEdges;
 - (void)setCells:(std::vector<sb::cell_t> const&)newCells;
-- (void)showMenu:(NSMenu*)menu withSelectedIndex:(NSUInteger)index forCellWithTag:(int)cellTag font:(NSFont*)font popup:(BOOL)isPopup;
+- (void)showMenu:(NSMenu*)menu withSelectedIndex:(NSUInteger)index forCellWithTag:(NSInteger)cellTag font:(NSFont*)font popup:(BOOL)isPopup;
 @end
 
 template <size_t cellCount> void SetCells (OakStatusBar* statusBar, sb::cell_t const (&cellList)[cellCount])
